+++ ICSharpCode.CodeConverter/CSharp/DeclarationNodeVisitor.cs
@@ -99,7 +99,7 @@ namespace ICSharpCode.CodeConverter.CSharp
                 .SelectAsync(async c => (UsingDirectiveSyntax) await c.Accept(TriviaConvertingVisitor));
             var usingDirectiveSyntax = usings
                 .Concat(_extraUsingDirectives.Select(u => SyntaxFactory.UsingDirective(SyntaxFactory.ParseName(u))))
-                .GroupBy(u => u.ToString())
+                .GroupBy(u => (Name: u.Name.ToString(), Alias: u.Alias))
                 .Select(g => g.First());
 
             return SyntaxFactory.CompilationUnit(