+++ Tests/CSharp/ExpressionTests.cs
@@ -2237,6 +2237,24 @@ End Class", @"public partial class Class1
         }
 
         [Fact]
+        public async Task GlobalNameIssue375()
+        {
+            await TestConversionVisualBasicToCSharpWithoutComments(@"Module Module1
+    Sub Main()
+        Dim x = Microsoft.VisualBasic.Timer
+    End Sub
+End Module", @"using Microsoft.VisualBasic;
+
+internal static partial class Module1
+{
+    public static void Main()
+    {
+        double x = DateAndTime.Timer;
+    }
+}");
+        }
+
+        [Fact]
         public async Task TernaryConversionIssue363()
         {
             await TestConversionVisualBasicToCSharpWithoutComments(@"Module Module1
@@ -2280,6 +2298,55 @@ End Class", @"public partial class Class1
     }
 }");
         }
+
+        [Fact]
+        public async Task AliasedImportsWithTypePromotionIssue401()
+        {
+            await TestConversionVisualBasicToCSharpWithoutComments(
+                @"Imports System.IO
+Imports SIO = System.IO
+Imports Microsoft.VisualBasic
+Imports VB = Microsoft.VisualBasic
+
+Public Class Test
+    Private aliased As String = VB.Left(""SomeText"", 1)
+    Private aliased2 As System.Delegate = New SIO.ErrorEventHandler(AddressOf OnError)
+
+    ' Make use of the non-aliased imports, but ensure there's a name clash that requires the aliases in the above case
+    Private Tr As String = NameOf(TextReader)
+    Private Strings As String = NameOf(VBCodeProvider)
+
+    Class ErrorEventHandler
+    End Class
+
+    Shared Sub OnError(s As Object, e As ErrorEventArgs)
+    End Sub
+End Class",
+                @"using System;
+using System.IO;
+using Microsoft.VisualBasic;
+using SIO = System.IO;
+using VB = Microsoft.VisualBasic;
+
+public partial class Test
+{
+    private string aliased = VB.Strings.Left(""SomeText"", 1);
+    private Delegate aliased2 = new SIO.ErrorEventHandler(OnError);
+
+    // Make use of the non-aliased imports, but ensure there's a name clash that requires the aliases in the above case
+    private string Tr = nameof(TextReader);
+    private string Strings = nameof(VBCodeProvider);
+
+    public partial class ErrorEventHandler
+    {
+    }
+
+    public static void OnError(object s, ErrorEventArgs e)
+    {
+    }
+}");
+        }
+
 
     }
 }